Fix bug: some times a photo will apper twice back on back.
And rewrite the items memory management in column

1. Fix bug:
The two same photo always back on back and in the same column, and the
firest one's photo is not compated with the view.

Reson of it is when adding new photo in a column, the column height is
still not change to pre. photo's bottom, this make the new photo have same
measured height with previews one. In this situation, the new photo will
override the previews one in photo table.

2. Rewrite memory management:
Using dirty queue and reload/recycle needed mark, to same time when the
photo is recycle and reload in the same time (this will happened when
scorlling very fast).